+++ markdown.d.ts
@@ -120,6 +120,8 @@ export namespace MarkdownParser {
   export interface InlineParser extends
     Markdown<'inline'>,
     Parser<HTMLElement | Text, [
+      InlineParser.ParenthesisParser,
+      InlineParser.AnnotationParser,
       InlineParser.InsertionParser,
       InlineParser.DeletionParser,
       InlineParser.StrongParser,
@@ -129,12 +131,25 @@ export namespace MarkdownParser {
       InlineParser.CodeParser,
       InlineParser.ImageParser,
       InlineParser.LinkParser,
-      InlineParser.AnnotationParser,
       InlineParser.HTMLParser,
       InlineParser.TextParser
     ]> {
   }
   export namespace InlineParser {
+    export interface ParenthesisParser extends
+      // (abc)
+      Markdown<'parenthesis'>,
+      Parser<HTMLElement | Text, [
+        InlineParser
+      ]> {
+    }
+    export interface AnnotationParser extends
+      // ((abc))
+      Markdown<'annotation'>,
+      Parser<HTMLElement, [
+        InlineParser
+      ]> {
+    }
     export interface InsertionParser extends
       // ++abc++
       Markdown<'insertion'>,
@@ -195,13 +210,6 @@ export namespace MarkdownParser {
       // [abc](url)
       Markdown<'link'>,
       Parser<HTMLAnchorElement, [
-        InlineParser
-      ]> {
-    }
-    export interface AnnotationParser extends
-      // ((abc))
-      Markdown<'annotation'>,
-      Parser<HTMLElement, [
         InlineParser
       ]> {
     }