+++ src/parser/inline/strong.test.ts
@@ -32,10 +32,7 @@ describe('Unit: parser/strong', () => {
     });
 
     it('triple', () => {
-      assert.deepStrictEqual(inspect(parser('***<ruby>***')), [['<strong><em>&lt;ruby&gt;</em></strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('***`<ruby>`***')), [['<strong><em><code>&lt;ruby&gt;</code></em></strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('***`<wbr>`***')), [['<strong><em><code>&lt;wbr&gt;</code></em></strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('***[](#)***')), [['<strong><em><a href="#">#</a></em></strong>'], '']);
+      assert.deepStrictEqual(inspect(parser('***a***')), [['<strong><em>a</em></strong>'], '']);
     });
 
   });